+++ serviceTest/CHANGELOG.md
@@ -19,6 +19,7 @@ Changed
 - Check if log type is available before trying to get the log
 - Add a few more helful chrome option
 - Supports scala 2.13
+- Target java 11
 
 Added
 - Rawls billing v2 support
@@ -28,7 +29,7 @@ Added
 - Added more Scalatest tag descriptors intended to target tests to specific environments
 - Added dataRepoApiUrl to CommonConfig
 
-SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-service-test" % "0.18-23f5ae7"`
+SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-service-test" % "0.18-TRAVIS-REPLACE-ME"`
 
 ## 0.17